refactor(halo-app): Improve window telemetry
