+++ packages/apps/halo-app/src/hooks/index.ts
@@ -3,4 +3,3 @@
 //
 
 export * from './useSafeSpaceKey';
-export * from './useTelemetry';
