+++ packages/apps/halo-app/src/telemetry/base-properties.ts
